Format changes

+++ fuvest/extract/extract.py
@@ -24,11 +24,12 @@ def extract_fuvest():
         extract_fuvest_year(dir, ids, merged_list)
 
     amostra = pd.concat(merged_list)
-    amostra = amostra.drop(columns=['nome', '_merge', 'ano_ingresso_curso'])
+    amostra = amostra.drop(columns=["nome", "_merge", "ano_ingresso_curso"])
     amostra = amostra.drop_duplicates()
-    amostra.id = amostra.id.astype('int64')
-    amostra.origem_cpf = amostra.origem_cpf.astype('int64')
+    amostra.id = amostra.id.astype("int64")
+    amostra.origem_cpf = amostra.origem_cpf.astype("int64")
     write_fuvest_amostra(amostra)
+
 
 def extract_fuvest_year(dir, ids, merged_list):
     year = dir.parts[-1][-4:]
